refresh data partially working (newly added tickets in .NET are shown)
.NET => modify ticket => save
java => refresh => changes not visible 
.NET => add ticket => modify ticket (still in .Net)
java => refresh => new ticket appears in tableview with latest modifications visible
.NET => modify same ticket
java => refresh => these changes are not visible

The findAll() method in de DAO class only recognizes newly added tickets in the DB and does not notice modified tickets.

A full reboot of the application solves the problem but is inefficient.